# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amrutam Telemedicine Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/doctors/": {
      "get": {
        "tags": [
          "doctors"
        ],
        "summary": "List Doctors",
        "operationId": "list_doctors_api_v1_doctors__get",
        "parameters": [
          {
            "name": "specialty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Specialty"
            }
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rating"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorPublicResponse"
                  },
                  "title": "Response List Doctors Api V1 Doctors  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/doctors/{doctor_id}/slots": {
      "get": {
        "tags": [
          "doctors"
        ],
        "summary": "List Slots",
        "operationId": "list_slots_api_v1_doctors__doctor_id__slots_get",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doctor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailabilitySlotResponse"
                  },
                  "title": "Response List Slots Api V1 Doctors  Doctor Id  Slots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailabilitySlotResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doctor_id",
          "start_time",
          "end_time",
          "status"
        ],
        "title": "AvailabilitySlotResponse"
      },
...
      "DoctorPublicResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "bio": {
            "type": "string",
            "title": "Bio"
          },
          "consultation_fee": {
            "type": "number",
            "title": "Consultation Fee"
          },
          "rating": {
            "type": "number",
            "title": "Rating"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "specialty",
          "bio",
          "consultation_fee",
          "rating"
        ],
        "title": "DoctorPublicResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}